Create BlockchainMessager.sol
Here we created a messanger smart contract in which only owner can add a new message to the contract but at the same time any one can read that message on the blockchain. Aside from that we can also count how many times the message was updated by the owner.